Create testFiles.txt
A listing of the files that I created that were part of my regular regression. The total number of executed scenarios was 430-ish. Execution took 3.5-3.75 hours but manual execution took twice that amount of time.

env value was set to either "development" or "staging" depending on where I was running these.