Create venueAdminRefundCancelAllReservations.spec.js
A script that was used to loop through the orders to refund & cancel them. This checked the payment type (credit, cash, group or split payments between cash & credit) then refunded accordingly and cancelled the orders completely. This was done to allow for re-using the products as a limited amount were available.